[routing] Use derivative crate.
This change replaces some manual implementations of Clone and Debug in
the routing library with use of the "derivative" crate, which works
around the incorrect bounds-calculation code in the normal derive
macro.

This shouldn't change the behavior of anything; it should just remove
yadda yadda code.

Change-Id: I8a14d5a8d53d40024cb1be83e7b5beb8202f97d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/547543
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>