[kernel][vm] Unify accessed harvesting
It is no longer neccessary to know anything about a page, in particular
whether it is in a particular kind of VMO or not, to update its
accessed time in the page queues. This change leverages that to remove
the accessed callback, directly calling MarkAccessed from the mmu code,
and then performing a single iteration over the arch aspace for both
page aging and page table reclamation.

Change-Id: I6bc8d6c513f40532bf1da2faa8602e0c2e0f5272
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533288
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>