[gn_sdk] Fix run_py.mako to work under Python3.
The template's #! was updated to refer to Python3 rather than Python2,
but the template content was not updated to be Python3 compatible,
seemingly because these tests are not run anywhere.

Fix the template's "print" syntax to be valid Python3.

Bug: 60187
Change-Id: I49895d20af0ea153acfd690d28a3f25b6b0c6380
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546081
Fuchsia-Auto-Submit: Wez . <wez@google.com>
Reviewed-by: Wez . <wez@google.com>
Reviewed-by: Misha Gridnev <gridman@google.com>
Reviewed-by: Randall Bosetti <rlb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>