[amlogic] Dump registers on RDMA stalls.
There is circumstantial evidence that the RDMA engine is stalling on
these devices, leading to "frozen" screens that get refreshed with old
content. In the event that we cannot make forward progress for 2
seconds, dump all the registers. This should give us a better idea as
to whether the device is being programmed incorrectly or the silicon
has a bug.

Bug: 189400236
Test: build and e2e on local device.
Change-Id: Iea288f1f99f8038258f89ffbc24ca8b70b64c699
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/547132
Fuchsia-Auto-Submit: Randall Bosetti <rlb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>