[zxdb] modify check_for_agent
With a forwarded port, the fact that the debug agent port is open
doesn't mean that the debug agent is running (because the port is always
open).

This CL modifies check_for_agent to be able to handle two case:
- only one debug agent on a local or remote device.
- one or several debug agents on a local device. In that case, if one
debug agent uses the port 2345, it must be launched first.

Bug: 78902
Change-Id: Id184bb5954768de6f873204c726060d8992d39fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/547901
Commit-Queue: Vincent Belliard <vbelliard@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>