[flatland] Use channel instead of eventpair for link tokens
We currently use eventpairs that can be duped which can mess up the
lifetime expectations. We explained this in the documentation, but
we should have a better of enforcing this. This CL enforces it by
using channel instead which cannot be duped. Events still work
through ObjectLinker to match with their endpoints.

Fixed: 79509
Change-Id: I2c5f75ccf312efc94fffb23aa3f92fa2c664d350
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546908
Reviewed-by: Emircan Uysaler <emircan@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Emircan Uysaler <emircan@google.com>
API-Review: John Bauman <jbauman@google.com>