[debugger] Eval DW_OP_piece for unspecified data.
DW_OP_piece can accept no data in which case the bytes of the output are
said to be undefined.

This writes 0's to the data. This allows these expressions to be
evaluated but will be misleading if the undefined data is displayed.
Tracking that data is issue 79547.

Fix a bug in DwarfExprEval to clear the result data when it is
reinitialized. This was triggered by the added test. It is not triggered
normally because we normally create a new object for every expression.

Add an upper bound on the size of a piece to prevent runaway memory usage
on corrupt data.

Bug: 79528
Change-Id: I9fe0d9b5226bc9e06523b8d7f39eeb0f2a31be39
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548221
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>