[pm] Avoid metadata monitor test flakes
The filesystem watcher library used by pm receives events when the
monitored metadata file is written to, but that event correlates to a
write() call, not a close() after all the data is written.  Outside of
tests, metadata is staged outside the repository and then moved into it,
so that race window should be short.  Within tests, the tests write to
the monitored path directly, sometimes causing the monitor to fail to
parse the partially written metadata.

A more correct fix to this issue would be to have an event to consume
when a file is modified and then closed by the writer.  This change
hopes to cheaply reduce the flake rate by adding a retry to the
metadata reader until a more proper fix can be implemented, or pm is
replaced with a different tool.

Test: `--test.count 100 --test.run TestMetadata` with the metadata
writer writing 1 byte per syscall.  Increased the retry count until
flakes stopped happening on my machine.

Fixed: 78682
Change-Id: Ibe30505104f65bc0be6ca420eab76ff6da5f7f20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548157
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Commit-Queue: Kevin Wells <kevinwells@google.com>