[fdio][zxio] Move zxio type to posix conversions into zxio
Both fdio and zxio are interested in the POSIX mode bits associated with
a given I/O object. fdio has to set the st_mode field for stat()
operations on a file descriptor based on the object's type and
abilities. zxio has to internally compute this value to populate the
mode field in fuchsia.io/NodeAttributes. This moves the type specific
logic for computing the mode into zxio and removes the fdio 'dir' type,
which existed only to compute a directory-specific mode value.

The fuchsia.io2 definition of node attributes is defined independently
of POSIX. If / when we move to this definition, we can move the type
translations from zxio up to fdio.

This also moves the definition of zxio_get_posix_mode to a dedicated
header file. This provides a pure C definition for callers that need it
(such as Rust bindgen) and makes it easier to track down users for the
eventual attributes cleanup.

Bug: 43267
Change-Id: I7191257693fe130afb3edc72c504e5b4a5ca3367
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541941
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>