[fssh] Moving info logging to debug
Info level logging appears on stdout, which is intermingled with the
output of commands run on the device via ssh making the parsing of the
result difficult.

This moves the message to the debug level, so it can be used by
users to diagnose an issue, but not interfere with normal use.

Change-Id: Iaa4853d3c051a05ef1347cb303a97bbe4569422c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548483
Reviewed-by: Omer Lev-ran <omerlevran@google.com>
Reviewed-by: Sean Cuff <seancuff@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>