[netstack] Record stats upon DHCP packet drops
The DHCP client silently discards frames which are deemed irrelevant.
This change adds logging and statistics to track each instance in
which a frame is so discarded. The motivation is to provide better
data for investigations when a device is unable to acquire an IP
address.

Bug:77810

Change-Id: Ibb9c9811b680f3a6298d73abdf74429755cf95d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542021
Commit-Queue: Nick Brown <nickbrow@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>