[fxfs] Some corruption fixes
There was a deadlock writing to the journal: it would try and take a
read lock on the journal handle, but something else could be holding a
write lock and then be waiting to write to the journal.  It probably
shouldn't have been trying to get a read lock (and that's something I'll
look into shortly), but it seemed better to just fix it so that flushing
the journal is done asynchronously which was an outstanding TODO anyway.

The next problem was that we were not discarding journal extents when
we processed the Discard message.

The last problem addressed here is that we were not correctly accounting
for flushed writes when building the checksum list.  The checksum
verification is designed to only work up to the last point we know was
flushed.

Change-Id: If3e30d81bd6cae573dc09bc08be0587fc425fc45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546681
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>