[blobfs][host] Check total block count when compressing
Update the host tools compression decision logic to match the logic on
the target device.

With the compact Merkle tree format even if compressing a blob doesn't
save a block it may save enough space to fit more of the Merkle tree and
save a block.

Change-Id: I6cff53986effe547c8bc6a35d58a1ee224d9acc5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545424
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>