[debugger] Add a buffer that indicates valid bytes
Use this buffer in the DwarfExprEval and ExprValue to represent
composite values that might be partially optimized out.

Currently, only the basic getters in ExprValue check the valid bits.
Additional checks of these (especially for structs which is the main
user-case) will be hooked up in a follow-up.

Bug: 79547
Change-Id: I47fc9ada0cd8ab04dd367e17141b42d770ae2bb8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548497
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>