[debugger] Handle optimized out values.
Check for data validity when reading out bytes of an ExprValue. Set
"optimized out" errors when bytes aren't marked valid.

The return type of ExprValue.data() was changed to the new TaggedData
type which forced an audit of all users of the data() function that
might need to be checking for validity.

Bug: 79547
Change-Id: I250bee6fb10e44aa79ba942c8f0fb891c9ae594a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548764
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>