[pkg-resolver] Improve blob fetch error reducer
When a package resolve encounters an error writing a content blob for a
package, pkg-cache may decide that the error was fatal and tear down the
Get() operation, which also closes all other concurrent blob writes for
that operation.  The pkg-resolver was simply returning the first blob
fetch error as the root cause of the resolve failure, but the
pkg-resolver may observe a cancelled fetch before the more useful error
that triggered the fetch to be cancelled, resulting in it sometimes
reporting a generic INTERNAL error for resolve requests that should have
failed with a more specific error.

This change modifies the package concurrent content blob fetch flow to
* wait for each blob fetch to succeed/fail with a cancelled error, or
* wait for any blob fetch to fail with a non-cancelled error
returning the more interesting error encountered, if any.

Fixed: 79559

Change-Id: I6a340091940b07003a176ce4b7885a5d4d6647ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548729
Commit-Queue: Kevin Wells <kevinwells@google.com>
Fuchsia-Auto-Submit: Kevin Wells <kevinwells@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>