[starnix] Add basic futex support
This CL contains support for the basic WAIT and WAKE futex operations.
We put the thread to sleep in a port, which will let us interrupt the
wait to deliver signals in the future.

This CL does not check the value for WAIT yet. That check is difficult
without atomic reads from the child process, but we can probably get
something basic working in a followup CL.

Bug: 74727
Change-Id: I0bdc44ce518a184a6b462eb0cf64ac6c0cdded0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548490
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>