[starnix] Add support for blocking I/O
This CL adds a mechanism by which read and write operations can block if
O_NONBLOCK is not set. They install an observer on the FsNode, which
gets notified when the node is readable (or writable). The threads sleep
in a port, which will let us interrupt the sleep to deliver signals in
the future.

Test: PipeTest.WriterSideCloses
Bug: 77376
Change-Id: I7ffb828e4efce0f763534e81bfa5ded7f6a66c63
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548492
Commit-Queue: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>