[build] Remove -fexperimental-relative-c++-abi-vtables flags
Relative vtables is enabled by default on Fuchsia, so we can remove
these flags that explicitly enable it.

Note that we will still need to keep `use_cxx_relative_vtables` because
some external libraries like libneural_network_models.so are built with
some BUILD.gn files from the fuchsia source tree and may want to opt-out
of this feature for now. Once we confirm there are no external users
that disable this, we can then remove the GN flag and the config.

This also changes the `compiler_relative_vtables` config to either
explicitly enable/disable RV when included, depending on the value of
`use_cxx_relative_vtables`. This should not be used anywhere in the
source tree.

Bug: 45314
Change-Id: Ia74532dfdf315ef821394ecfcfcdd289c5f42da3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548141
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>