[audio_core] Rename ReadableStream::Buffer::gain_db
The current name is confusing: the term `gain_db` sounds like it should
represent gain that should be applied to the stream in the future. In fact,
it's the amount of gain that has *already* been applied to the stream.

Hence I renamed this field to `total_applied_gain_db`.

Change-Id: I2f25784d2c2f915b6493ce768e36a9b4fd4589c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544501
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>