[fvm] Log FVM info on startup.
Enhances FVM logging on startup to include more information. It now
includes the full header information in addition to the free slices, as
well as all partition information.

The previous code attempted to log the partition information but instead
logged the pointer because of a coding bug.

The GetFvmLocked() was renamed GetHeaderLocked() to match GetHeader().
This tripped me up while writing this patch.

Change-Id: I628a23032450d605aa2e8e13d1f3fcbc76cc51d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549903
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Chris Suter <csuter@google.com>