[virtualization] Add libvulkan.so.1 to romfs
Normally on Linux libvulkan.so.1 is the real name (e.g. soname) of the
vulkan loader, while libvulkan.so can be a symlink to the latest
version. We can use a variable from Vulkan-Loader's arg.gni file to
determine whether we should package libvulkan.so.1 (the correct name)
instead of libvulkan.so

All the executables we're packaging will use libvulkan.so.1 as the
soname, so we don't need to package or symlink libvulkan.so separately.

Bug: 79706

Change-Id: I460d20c1f88b9e22043fc801f25567a3851c8266
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548602
Reviewed-by: Craig Stout <cstout@google.com>
Reviewed-by: David Reveman <reveman@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: John Bauman <jbauman@google.com>