[debugger] Improve containing function search.
Adds a flag to the CodeBlock::GetContainingFunction() function to
control whether inlines are accepted or not. This is not currently used
but will be used in followup.

To support this, allow going to the enclosing function above an inline.
Adds a new CodeBlock::GetContainingBlock() function to abstract the
confusing differences between parent() and containing_block() for the
different types of blocks.

Bug: 6322
Change-Id: I63de235902e96577b3692193057f186769f1f960
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548943
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>