[botanist] Clarify "early stop" IP resolution behavior
I originally thought the "early stop" behavior was basically a no-op,
largely because the variable name `mDNSTimeout` made it seem like we
only allotted that amount of time to resolve an IP address before
exiting. Actually, `mDNSTimeout` only applies to each question packet,
not the entire IP resolution operation.

So I renamed `mDNSTimeout` to `mDNSQuestionInterval` and used the
existing `t.C` channel to implement the "early stop" functionality
rather than creating a separate channel, for simplicity.

So, after resolving the first IP address, we'll still only send one more
question to try to get the second IP address. But now we'll wait 2
seconds for a response to the final question before giving up, whereas
previously we would only wait for 1 second. This is not expected to have
a significant effect on the behavior of botanist.

Change-Id: I89b26bb50cfc1bccbb4f05317f93d304ed461de3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549383
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>