[fdio][zxio] Add a zxio_borrow operation and use for remotes and sockets
This adds an operation to "borrow" the primary handle for a zxio object
and wires it up for sockets and the remotes family. This operation is
inherently dangerous but is useful for interacting with an object in
combination with external information about how it is being used, such
as calling fdio_unsafe_borrow_channel while holding an fdio_t reference.

This also removes some knowledge of zxio object's internal knowledge
from fdio.

Bug: 43267
Change-Id: I38f8e948e7e6ca219cd7ea9241aa0cca76a68081
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549888
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>