[debugger] Consider CodeBlocks w/ no range empty.
Previously the CodeBlock implementation considered blocks with
unspecified code ranges as covering all instructions. As described in
the bug, this causes problems and seems contrary to the specification.
Instead, consider these empty.

In the sym-info implementation, get the FindNameContext from the frame
if there is one. This will prevent divergence in the behavior. The code
was written the way it was because the FindNameContext didn't used to be
exposed.

Bug: 79797
Change-Id: Idb6b24124b24dc572f20e512409f22e1d89ffeb9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549403
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>