Revert "[assembly] Do not supply fvm max size if not set"
This reverts commit 11dcd4b6bfa6cb1289e93b361e2dc26ee6f9d804.

This change broke clients, because their FVM could not expand post-boot.

The FVM has a set amount of space allocated for metadata, which directly
correlates to the maximum size the FVM can expand. The amount of metadata
is set at assembly time, and cannot change. If `--max-disk-size` is not
set at assembly time, then the FVM is assembled with a minimum space
reserved for metadata, which means that the FVM cannot expand at all.

This change removed the `--max-disk-size` flag for clients that did not
supply it explicitly, hiding the fact that clients _should_ have been
supplying it.

Bug: b/192085906
Change-Id: I6e67b62abd64618cca3774f9b424c5de737f6d56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549885
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Aidan Wolter <awolter@google.com>