[gidl] Use gidlir.Value type consistently
This CL replaces many uses of interface{} with gidlir.Value. The latter
is a wrapper around interface{} with the possible types it can contain
documented in value.go. It was only used in a few places before, and not
in others for no apparent reason.

Test: fx test gidl conformance
Change-Id: Ib677f01bc053957bbcb3031068dc8c9965fddd67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550161
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>