[network-device] Migrate tests to WireSharedClient
fidl::Client makes an upcall to user objects to indicate the binding has
been torn down. This behavior isn't a good fit when the user object owns
the client, since it easily leads to use-after-free of user objects.

We are splitting fidl::Client to two separate client types that are more
specialized to different threading needs, in doing so deprecating
fidl::Client.

fidl::WireSharedClient is a client type that is designed to own
the user object. By inverting the ownership relationship, it is possible
to move or clone a WireSharedClient to a different thread, with the
user object managed through a two-phase shutdown pattern. It is suitable
for multi-threaded or functional usage styles.

See //zircon/system/ulib/fidl/include/lib/fidl/llcpp/client.h for more
details.

TEST: fx test -od netdev-client-cpp-test
Change-Id: Ie155d0333f6d092054844ad9831ee704f2c7525a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550721
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>