[kernel][hypervisor][arm64] Reduce EL2 address space size.
EL2 currently has an address space with the same shape as userspace. In
practice, EL2 doesn't need as much address space as userspace, as it is
just a 1:1 mapping from virt to phys.

Update EL2 to use the "IDENT" address space shape (currently defined as
39 bits). In the process, we rename some of the constants used to define
the shape of the identity page tables to match the GUEST/USER/KERNEL
constant names.

Bug: 79743
Change-Id: I3ad929413760801184538f085fe6d039b5cf933a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549721
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>