[bt][a2dp] Remove "eager" designation for component
Because bt-a2dp is included in the "universe" set in some
configurations, we run into issues in system startup due to issues
described in fxbug.dev/78801. As a temporary measure, remove the "eager"
tag for the component - this will not be an issue as the component will
be started when a client requests any A2DP-providing capabilities.

Bug: 79964
Test: Manifest only, fx build (this core_shard is currently not used)

Change-Id: Idef7f15330a3c18fe32017d652bb2d492da2ffb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550401
Reviewed-by: Justin Mattson <jmatt@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>