[bt][manifest-smoke-test] Fix flake in bt-init smoke test
Serve a pseudo-directory to bt-init (and thus bt-gap) so that bt-gap
does not panic when it tries to open the bt-host device directory.

Also, wait for realm destruction prior to ending the test, to ensure
the mock component running the pseudo-directory terminates after the
bt-gap component in the realm depending on it.

MULTIPLY: bt-init-smoke-test (fuchsia): 150
Fixed: 77775
Test: First, apply fxrev.dev/541128 and verify that this patch reliably
      reproduces the flake seen in fxbug.dev/77775. Then, apply this
      patchset and verify that the flake is fixed.
Change-Id: I96d94d957cda273861724ff2db36513e3363b3e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541129
Commit-Queue: Lucas Jenkins <lucasjenkins@google.com>
Fuchsia-Auto-Submit: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>