[fshost] Set minfs-max-bytes correctly.
When minfs-max-bytes is not needed for a configuration, it should be
omitted from the fshost configuration. Currently it's being set to the
empty string which is giving errors in the syslog at boot.

The issue was that the name of the argument from the invoker in the
fshost config template was actually catching the build parameter. The
result was the default configuration looked like it was overriding the
value with "". This change just uses a different name for the invoker's
override parameter.

There should be no behavior change since this set properly when needed,
and it was ignored when set to empty. This just removes a stray warning
in the syslog.

Bug: 77049
Change-Id: Ia960dd7e76afccb96ff16fc91a67c5bac1ec2d90
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549901
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Chris Suter <csuter@google.com>