[brcmfmac] Fix tracking of disconnect origin
The driver currently tracks the origin and type of disconnects to
determine what kind of notification to send to SME when the disconnect
has completed. The variable to track was not properly reset in one place
leading the driver to sometimes think that any disconnect event from
firmware was actually the result of a disconnect initiated by SME. This
in turn resulted in the driver sending a confirm event rather than an
indication event to SME. SME then ignored this event because it made no
sense and thus SME was not aware that the device had been disconnected
from the network and so of course no attempts to reconnect were
initiated.

Fix this by clearing the tracking variable in the one place it was
missing. Also added a sim test to catch this issue.

Bug: b/191086331
Test: brcmfmac_sim_test, wlan_tests, added sim test
Change-Id: I83f918e1ed7824c77b05a50a920a692850372a27
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544828
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
(cherry picked from commit 870008719ba5def3b857bcc584c0de4db5a10438)