[diagnostics][tests] Fix flake in cpu stats
It turns out that appmgr can have a measurement taken when it hasn't
been in the queue at all. This is likely because it is one of the only
things running at system startup, so it doesn't need to wait for CPU
time in the scheduler queue.

This CL fixes this test where it would fail due to a 0 queue time for
appmgr.

Fixed: 80020
Change-Id: Icb8fac606d55863b3e04e4a0dab70924ebbcd008
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551161
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>