[netstack] Fix test_case usage in DHCP
The `test_case` macro doesn't assert on a Result returned from the test
body unless the result is specified with `=> foo`. Since we
test_case-ified the DHCP tests with a Result return value, they have
been silently passing.

This change modifies the tests to not return a result and also fixes a
bug in the tests that was masked by this behavior.

Updates: 73252
Change-Id: I77ed459bb619128f4722f28f07942e729fed04bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550604
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Nick Brown <nickbrow@google.com>