[rbe] reject absolute path deps from remote build
Dep paths like /b/f/w/... only make sense in a remote context,
and should never appear in local depfiles.
Relative paths are required for remote-executable actions.

This caught a problem in a third_party dependency on grammar.pest.

There is a temporary workaround to modify the remote-built depfile,
while we wait for the fix to land (unblock for now).

Change-Id: I4c167a0975e94655235b4c44330f7076e6d8b550
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550041
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>