[bt][lib] Add Vigil helper library
Vigil is a library which lets you add actions that are queued to happen
when the item is dropped. It defines a trait, DropWatch, which takes a
function to be called when the item has been dropped.

To be used in A2DP and HFP profiles.

Bug: 64521
Tests: added unit tests; fx test vigil-tests

Change-Id: I080f810295d16972f211548de9d60e674cee622a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548601
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>