[test][docs] Update v2 test examples
Now that RealmBuilder is available, add a reference to the RealmBuilder
page.
This also replaces the example that used the driver pattern as Archivist
no longer needs to be manually placed in a test realm. Instead, this
adds a simple example where the test connects to a sibling component.

Change-Id: Ibe0d44137d1983e98443459d82c3a87093b74fa8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/547904
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>