[libsync] don't remove sanitizers config.
This CL modifies the libasync library so that it can be build with
different sanitizers. In the past, that configuration was removed
because it would cause trouble during program startup.

It seems like at some point code got refactored and this is no longer
blocking program startup.

Bug: 73210
Change-Id: Iaaac20505da2592439c4d160e9326b4f270090fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507383
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>