[netstack] Support route preferences
This change adds support for routes to be sorted based on preference
and for the NDP dispatcher to add routes with some preference value.

Routes to the same destination will be sorted based on their preference.
If the destination and preference are the same, then the metric value is
used to break the tie (then following the existing sorting algorithm for
routes).

Bug: 79015

Change-Id: Ic6f4a94f18c7b747ad9f038f333dcedbf20a09f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548354
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>