[ffx] Handle non-existent files in the sdk manifest.
Make the File handle Optional, and if it doesn't exist return empty
data. Previous behavior is that if some file in the manifes.json is
missing, then the sdk plugin fails. We should want the parser to be more
resilient, instead of erroring, we'll set empty data to the manifest
field.

Change-Id: I551e7f18d23272d2234354cb06f308a462bc290b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551380
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Andy Weiss <dragonbear@google.com>
Commit-Queue: Yuan Zhi <yuanzhi@google.com>