[zircon] Add ZX_EXCEPTION_STATE_THREAD_EXIT
Currently, userspace can either handle an exception or have the
exception continue to propagate. This CL adds the ability for userspace
to make an exception cause the thread to exit.

This mechanism will be used by starnix to terminate threads in the child
process. When a thread in the child process calls the Linux exit() UAPI,
Starnix catches the BAD_SYSCALL exception and marks the exception as
FATAL, which causes Zircon to tear down the underlying Zircon thread in
the child process.

Test: Threads.ExitViaException
Bug: 77376
Change-Id: I3120a7287a6e99e215c547e0ff955b288044c71f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550383
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Commit-Queue: Adam Barth <abarth@google.com>