[fx repro] Read fint_params_path from properties
This is more robust that constructing the path using the builder
project/bucket/name, which may break in the future if, for example, we
ever have multiple builders use the same fint params file. The existing
strategy already doesn't work for subbuilds and rerun builders, which
have names that differ from the fint params files they use.

This relies on recipes change fxrev.dev/551763, but is backwards
compatible with older builds. We'll probably want to keep the backwards
compatibility around for a year or so.

Bug: 80137
Change-Id: I9bcdd8581a4d8521c462bf8f16eb49be8d7224d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551819
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>