[zbi] Fix compression control flow
It turns out that ZSTD_compressBound() does not actually return an
upper bound on the size of the would-be compressed data: its
documentation reads
```
When there are a lot of small blocks due to frequent flush in streaming
mode the overhead of headers can make the compressed data to be larger
than the return value of ZSTD_compressBound().
```

The compression logic previously assumed that an output buffer of that
size would be sufficient - and a counterexample was met in the wild,
resulting in a crash. In this change, we update that logic to
iteratively request more output buffers as they are filled.

Change-Id: I656cfd57f13496829b831a4f11363665f7f863a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/551887
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Ashley Coleman <ashleycoleman@google.com>